+++ tests/Test/ConstrainedPlacementNoStackingTestItem.php
@@ -23,12 +23,12 @@ class ConstrainedPlacementNoStackingTestItem extends TestItem implements Constra
      */
     public function canBePacked(
         PackedBox $packedBox,
-        int $proposedX,
-        int $proposedY,
-        int $proposedZ,
-        int $width,
-        int $length,
-        int $depth
+        float $proposedX,
+        float $proposedY,
+        float $proposedZ,
+        float $width,
+        float $length,
+        float $depth
     ): bool {
         $alreadyPackedType = array_filter(
             iterator_to_array($packedBox->getItems(), false),
